Adding SpaMo library reqs.

+++ spamo/t5_slt.py
@@ -164,7 +164,7 @@ class FlanT5SLT(AbstractSLT):
         self.t5_model = T5ForConditionalGeneration.from_pretrained(
             t5_model, 
             cache_dir=self.cache_dir,
-            torch_dtype=torch.bfloat16,
+            torch_dtype=torch.float32,
             use_safetensors=True 
         )
